requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
